refactor(app): refactor for beta.12
