--- app/pages/hello-ionic/hello-ionic.ts
@@ -1,11 +1,0 @@
-import {Component} from '@angular/core';
-
-
-@Component({
-  templateUrl: 'build/pages/hello-ionic/hello-ionic.html'
-})
-export class HelloIonicPage {
-  constructor() {
-
-  }
-}
