--- app/pages/item-details/item-details.scss
@@ -1,7 +1,0 @@
-div.selection {
-  font-size: 22px;
-  display: block;
-  text-align: center;
-  padding-top: 5%;
-}
-
