--- app/pages/item-details/item-details.ts
@@ -1,15 +1,0 @@
-import {Component} from '@angular/core';
-import {NavController, NavParams} from 'ionic-angular';
-
-
-@Component({
-  templateUrl: 'build/pages/item-details/item-details.html'
-})
-export class ItemDetailsPage {
-  selectedItem: any;
-
-  constructor(public navCtrl: NavController, navParams: NavParams) {
-    // If we navigated to this page, we will have an item available as a nav param
-    this.selectedItem = navParams.get('item');
-  }
-}
