--- app/theme/app.core.scss
@@ -1,12 +1,0 @@
-// http://ionicframework.com/docs/v2/theming/
-
-
-// App Shared Imports
-// --------------------------------------------------
-// These are the imports which make up the design of this app.
-// By default each design mode includes these shared imports.
-// App Shared Sass variables belong in app.variables.scss.
-
-@import "../pages/hello-ionic/hello-ionic";
-
-@import "../pages/item-details/item-details";
