+++ src/app/app.component.ts
@@ -1,12 +1,15 @@
-import {Component, ViewChild} from '@angular/core';
-import {ionicBootstrap, Platform, MenuController, Nav} from 'ionic-angular';
-import {StatusBar} from 'ionic-native';
-import {HelloIonicPage} from './pages/hello-ionic/hello-ionic';
-import {ListPage} from './pages/list/list';
+import { Component, ViewChild } from '@angular/core';
+
+import { ionicBootstrap, Platform, MenuController, Nav } from 'ionic-angular';
+
+import { StatusBar } from 'ionic-native';
+
+import { HelloIonicPage } from './pages/hello-ionic/hello-ionic';
+import { ListPage } from './pages/list/list';
 
 
 @Component({
-  templateUrl: 'build/app.html'
+  templateUrl: 'app.html'
 })
 class MyApp {
   @ViewChild(Nav) nav: Nav;
@@ -32,7 +35,7 @@ class MyApp {
     this.platform.ready().then(() => {
       // Okay, so the platform is ready and our plugins are available.
       // Here you can do any higher level native things you might need.
-      StatusBar.styleDefault();
+      // StatusBar.styleDefault();
     });
   }
 
@@ -43,5 +46,3 @@ class MyApp {
     this.nav.setRoot(page.component);
   }
 }
-
-ionicBootstrap(MyApp);
