+++ src/app/app.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core';
+import { IonicApp, IonicModule } from 'ionic-angular';
+import { MyApp } from './app.component';
+import { HelloIonicPage } from '../pages/hello-ionic/hello-ionic';
+import { ItemDetailsPage } from '../pages/item-details/item-details';
+import { ListPage } from '../pages/list/list';
+
+@NgModule({
+  declarations: [
+    MyApp,
+    HelloIonicPage,
+    ItemDetailsPage,
+    ListPage
+  ],
+  imports: [
+    IonicModule.forRoot(MyApp)
+  ],
+  bootstrap: [IonicApp],
+  entryComponents: [
+    MyApp,
+    HelloIonicPage,
+    ItemDetailsPage,
+    ListPage
+  ],
+  providers: []
+})
+export class AppModule {}
