+++ src/pages/hello-ionic/hello-ionic.html
@@ -1,6 +1,6 @@
 <ion-header>
   <ion-navbar>
-    <button menuToggle>
+    <button ion-button menuToggle>
       <ion-icon name="menu"></ion-icon>
     </button>
     <ion-title>Hello Ionic</ion-title>
@@ -8,7 +8,7 @@
 </ion-header>
 
 
-<ion-content padding class="getting-started">
+<ion-content padding>
 
   <h3>Welcome to your first Ionic app!</h3>
 
@@ -19,7 +19,7 @@
     Follow along on the tutorial section of the Ionic docs!
   </p>
   <p>
-    <button primary menuToggle>Toggle Menu</button>
+    <button ion-button color="primary"" menuToggle>Toggle Menu</button>
   </p>
 
 </ion-content>
