+++ src/pages/hello-ionic/hello-ionic.scss
@@ -1,4 +1,4 @@
-.getting-started {
+.hello-ionic-page {
 
   p {
     margin: 20px 0;
