+++ src/pages/hello-ionic/hello-ionic.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+
+@Component({
+  templateUrl: 'hello-ionic.html'
+})
+export class HelloIonicPage {
+  constructor() {
+
+  }
+}
