+++ src/pages/item-details/item-details.scss
@@ -1,0 +1,11 @@
+.item-details-page {
+
+  div.selection {
+    font-size: 22px;
+    display: block;
+    text-align: center;
+    padding-top: 5%;
+  }
+
+}
+
