+++ src/pages/item-details/item-details.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+
+import { NavController, NavParams } from 'ionic-angular';
+
+
+@Component({
+  templateUrl: 'item-details.html'
+})
+export class ItemDetailsPage {
+  selectedItem: any;
+
+  constructor(public navCtrl: NavController, public navParams: NavParams) {
+    // If we navigated to this page, we will have an item available as a nav param
+    this.selectedItem = navParams.get('item');
+  }
+}
