+++ src/pages/list/list.scss
@@ -1,0 +1,3 @@
+.list-page {
+
+}
