--- www/service-worker.js
@@ -1,11 +1,0 @@
-self.addEventListener('activate', function (event) {
-
-});
-
-self.addEventListener('fetch', function (event) {
-
-});
-
-self.addEventListener('push', function (event) {
-
-});
